Rate limit by Cloudflares "CF-Connecting-IP" header when possible
Django Rest Framework's default configuration uses the X-Forwarded-For
header to identify clients.  The presence of Cloudflare in our
production/staging system complicates this, and from testing a little
bit I don't think we should rely on X-Forwarded-For.

But we can obtain the IP address that the website visitor connected to
Cloudflare with from the CF-Connecting-IP header.

More reading:

https://support.cloudflare.com/hc/en-us/articles/200170986-How-does-CloudFlare-handle-HTTP-Request-headers-

https://www.django-rest-framework.org/api-guide/throttling/#how-clients-are-identified