api: Add pagination links by default, optional envelope
Updates the incident list API endpoint to provide next and previous
page links in the `Link` header.  This is the pagination format that
will be used by default for requests into this endpoint.

Also adds an optional `envelope` argument, which can be sent with the
`envelope=1` query parameter.  If this is sent, then the pagination
links will be included inline with the response, alongside the actual
results in the `results` field.  This behavior is Django Rest
Framework's built-in pagination style, I am choosing to make it
opt-in.