Rework common app's test factories to be non-random
I've moved the "randomized" data factories into their own file:
devdata.py.  For the test factories, I've changed them around not to
depend on any Faker-based properties.

The reason for this is I was running into some local test suite
failures based on data randomization.  Even though we are now using
the same seed for all test runs, changing the order or number of
tests can still affect the subsequent randomized data that's
generated.  One solution to this is what I've started here (limited to
the "common" app) and another is resetting the seed at the beginning
of each test.  I prefer the first option.