Experimentally diff coverage analysis to Circle CI
This will add a new job to our pull request CI suite that checks the
difference in test coverage between the PR branch and the develop
branch.  The goal here is to add more visibility to the coverage
changes as they happen.  I'm not completely set on the
`--fail-under=100` option, but it's what I'm starting with to see if
it works.