Rework ManyRelationFilter to allow filtering by text fields
This involved some fairly extensive modification of the code relating
to this filter, in particular the functioning of the categories filter
is used in several places *other* than the actual filtering of the
incidents (e.g. in the summary, on the incident index page).  So that
had to be modified.

However, what's going on here is a mostly straightforward conversion
of the filter from using a list of integers to a pair of lists: either
a list of integers or a list of strings.  I've used a dataclass for
this new object, so that it's more explicitly defined than, say, a
dictionary or a tuple.

One thing that's not done here is applying the `text_fields` argument
to all possible `ManyRelationFilter` instances.  That will come later
and I'm going to write some tests for that.  What this commit is doing
is just making it possible to filter on text values and then following
out the repercussions of that change to the places where the tests
were failing or errors were happening in subclasses of
`ManyRelationFilter`.