Update jest version to 27
This updates the versions of the packages jest and babel-jest to the
latest in the 27.x line.

I've also updated the jest configuration so that the tests pass.  This
means setting the `testEnvironment` configuration option to
`jsdom` (default is `node`).  One side effect of this, I discovered,
was that `window.setImmediate` was removed.  I think
`process.nextTick` is an appropriate substitute for what we're using
it for here, even if it's not identical.