incidents: Rework `recently_updated` method to check cache
Since the `_incident.html` template is used in several places, and
we're not sure exactly what the state of the incident model will be
going into it, we cannot be sure that `updated_days_ago` will be
populated.  This property is not a model attribute but an annotation
that has to be applied at the query level.  In particular, it will
_not_ be set if the incident being rendered is a featured page on the
home page.

I've solved this problem by using the model method `recently_updated`
to either defer to `updated_days_ago` if present, or query the
database to find the value if it's not present.  This should correctly
deduce if an incident was recently updated in all cases, while
preserving efficiency where possible.

One other benefit is we get to remove a "magic number" from the
template code.