Add a new charts app
This app will hold all of our logic around the forthcoming charts
in the new tracker design. I've also added a corresponding
javascript entry point and updated the webpack config to compile it.

For now this app will consist of a single view that renders sample
charts from the javascript.