openapi parameters: Prevent crash in CI environment
On CI, or in any case where the database has not been fully loaded and
migrated, calling this function will result in an error (e.g. "table
does not exist").  This can and does happen with this function as it's
used in a decorator on an API view, and I've observed that it gets
evaluated before migrations are loaded when Django is starting up.

This is kind of a kludge, but it did prevent the error for me locally,
while still returning the correct results _after_ the DB loaded.