Use dev-data factories to render Incident Cards on the styleguide
This change required a small expansion of what's going on in the style
guide view.  Mostly, we are assembling some dev data in-memory (both
to make it fast and so that nothign improper happens if we load this
page on staging/production) so that the incident card template will
have something to render in the styleguide.

I will note: we apparently can't use the factory functions as
explicitly as I'd like.  When using the build strategy, the relations
weren't appearing for me (possibly related to this issue
https://github.com/FactoryBoy/factory_boy/issues/856), but we have a
workaround in the form of Modelcluster (similar to how wagtail creates
previews).  We can create instances of the related objects (using the
factory build strategy), but we have to "relate" them manually.  I
think this is fine, though it's more boilerplate than is ideal.