Take changes in category symbols branch into account
This is actually great since we are no longer tied to the `text-link`
class and the extra svg declarations we had to use to apply the icons
to buttons are no longer needed.

Also, I created the categories for each icon on the styleguide page so
we're not reliant on using the database.