Account for date fields having two form inputs
I am not in love with the `continue` statement here, and I'm sure
there is a slightly nicer way to do this, but I'm reluctant to make
too many changes to rest of the code here, and this felt like the
least alteration I could make that would accomplish the goal.

Very happy to hear some other ideas about how this could be done
better in the future (e.g. convert `field` to a list).