Align database card columns with page grid
There's a lot going on in this commit, as I restructured a lot of
things, which broke some other things (including the "Show more/less"
button, which may have been partially broken this whole time) so I
took the time to go through and clean up a lot of stuff in addition to
aligning the columns.

Note: At "narrow desktop" widths, we are running into longish columns
overflowing their boundaries.  Probably should fix this.