Add 3px buffer to homepage actions to contain focus styles
Without this, the focus outline is partially hidden on the bottom
edge.  This, I think, is because `.homepage-section-inner` has
`overflow-x: hidden`, which actually causes `overflow-y` to change
from `visible` (the default) to `auto`.

If there's a better way to handle this, please let me know because I
am not in love with this approach.