Rework incident card details table to match design
The design of the home page incident cards has the details spaced more
according to what the incident details are on the "database card"
template.  This brings the "incident card" into better conformance
with the design by borrowing those styles, though possibly that group
of CSS classes should be renamed.