Fix boolean field form filter
"Yes" and "No" are not actually valid things to submit to our incident
filter code as a boolean value.  The value is processed by the
`to_python` method on the field, which accepts "0", "f" and "False"
for the `False` value (and similarly for `True`).  0 and 1 seem like
fine things to put in the filter form to me.