Ignore latest vulernability in `minimist` package
It is not looking likely to me that we will be able to update all of
our code that depends on minimist in a timely manner.  Here's the
dependency output of npm audit:

```
node_modules/minimist
  gonzales-pe  >=3.0.0-1
  Depends on vulnerable versions of minimist
  node_modules/gonzales-pe
    postcss-sass  *
    Depends on vulnerable versions of gonzales-pe
    node_modules/postcss-sass
  json5  1.0.0-beta - 2.2.0
  Depends on vulnerable versions of minimist
  node_modules/json5
  node_modules/loader-utils/node_modules/json5
  node_modules/tsconfig-paths/node_modules/json5
    loader-utils  1.2.0 - 1.4.0
    Depends on vulnerable versions of json5
    node_modules/loader-utils
      babel-loader  >=8.1.0
      Depends on vulnerable versions of loader-utils
      node_modules/babel-loader
      sass-loader  8.0.0 - 8.0.2
      Depends on vulnerable versions of loader-utils
      node_modules/sass-loader
    tsconfig-paths  >=3.3.0
    Depends on vulnerable versions of json5
    Depends on vulnerable versions of minimist
    node_modules/tsconfig-paths
      eslint-plugin-import  >=2.21.0
      Depends on vulnerable versions of tsconfig-paths
      node_modules/eslint-plugin-import
        eslint-config-airbnb  >=18.2.0
        Depends on vulnerable versions of eslint-plugin-import
        node_modules/eslint-config-airbnb
        eslint-config-airbnb-base  >=14.2.0
        Depends on vulnerable versions of eslint-plugin-import
        node_modules/eslint-config-airbnb-base
  mkdirp  0.4.1 - 0.5.5
  Depends on vulnerable versions of minimist
  node_modules/mkdirp
    modernizr  <=3.8.0 || >=3.9.1
    Depends on vulnerable versions of mkdirp
    node_modules/modernizr
```

So `modernizr`, `eslint-config-airbnb{-base}`, `sass-loader`,
`babel-loader`, and `postcss-sass` all depend eventually on
`minimist`.  As of this writing, there has not been a new release of
`minimist` (the last of which was two years ago), so either that
package would have to release a fixed version (and many of the above
transitive dependencies also release new versions) or those same
dependencies would have to release versions not using `minimist` at
all.

Given the number of packages concerned here, I don't think this is
going to happen soon enough for this failed check not to start
annoying us.  So I'm adding this vulnerability to the ignore list.