eslint: Ignore the charts directory JS
There are 200+ errors in this directory, and we know about them
already.  Let's ignore it on CI until we can fix them so that we don't
see failed builds everywhere.

Also fix the stylelint configuration for ignoring files.  I think the
old code was not ignoring anything, the documentation says that the
key to use is `ignoreFiles`:

https://stylelint.io/user-guide/configure/#ignorefiles