Remove wagtail's legacy richtext setting
This commit removes the setting that wraps all content rendered via
the `richtext` filter with a `<div class="rich-text">` tag.  We were
using that `rich-text` class to apply consistent styles to all of our
rich text content.  However, there are places where we need the
styling to be different, so I've gone to each place where the
`richtext` filter is used and either:

1. Hard coded the tag (or CSS class) in that same place
2. Added Sass that corrects the styles for that piece of content, and
applied the needed bits from the `.rich-text` styles.

I like this approach because the "legacy rich text" wagtail setting
did create some confusion by magically adding that div, and I think
this makes adding it into something we explicitly have to want for
that piece of text.