Convert all date calculations to UTC for consistency
We can't predict the time zone of the user's browser and it wouldn't
even give us much benefit to try. The closest thing to using timezone-naive
dates in javascript is keeping all dates in UTC

In the process I've also rewritten bits of the react components to
expect date objects rather than strings. We now perform parsing ahead
of time in in the data loader