Ensure chart redirects to filter with YYYY-MM-DD format
This format is required by the `<input type="date">` fields in the
sidebar, and while the backend filter properly parses the date, the
value supplied to the form will not be understood or populated in that
field when the page renders.  This PR fixes that.

I've also fixed what I think is a logic bug in the "six month" date
range.  The previous version of this code did not include the current
month in the filter range, even though the chart includes that month.
I've changed that code so that the database page is filtered to
include the only same six months that are shown on the right-side bar
graph.