info table: Display image on hover for desktop screen sizes
This is a little jank, I'm going to be honest.  I'm not sure I know
how to do it, otherwise, and also I'm not sure how to get the reverse
of the fade in.  Having the containing box become `display: none` when
the parent is not hovered makes it vanish instantly.