Rename "Physical Attack" category to "Assault"
The slug for this category page is changing, so we're moving
everything we can here from the old name to the new name.  Not
everything changed in this commit is required for the site to still
work, but it makes it easier to change as much as we can in one go.