Add prop for tooltip (when unselectable) to our `Button` components
This might be kind of a hack, but I couldn't really figure out how
else to display this!  But on the other hand I also feel like it's
just adding some flexibility to our "Button" and "ButtonsRow"
components even if we're not using it anywhere but one specific set of
buttons just yet.