Improve how we determine headers in `get_renderer_context`
Relying on the path name is going to end up being unreliable, but the
action name is set on the viewset class by Django Rest Framework, so
we can check that instead.  This is still a string, which is slightly
awkward, but at least it's the internal name of our own action and not
based on any kind of user input or other external interface.