incident category details: Always display category headers as links
In category tables (which appear on the database page and incident
detail page), always render the category headers as links to the
appropriate category page.  This means that the "Chilling Statement"
and "Other Incident" categories will be shown as links without
corresponding lists of field values (since there are no incident
fields belonginging to those categories).