category field values: Sort metadata-less categories last
This updates `get_category_details` to insert categories without field
data into the return value last, so that when we iterate through this
result with `.items()` they will be retrieved last.  Ultimately, this
is because we want to display these categories in the UI _after_ the
ones that have metadata fields associated with them.