Add initial scaffolding for filter sidebar react component
1. Adds a separate entry point in webpack, along with a seprate top
level component called `FilterSidebar`.  This is to preserve (as much
as is feasible) the behavior of the app at `/charts/`.  Though some of
this has to change unless we want to copy every existing component,
which I'm not sure we really do.

2. Load the `FilterSidebar` on the incident index page and pass in
initial props containing "filter definitions" (which is what I'm
calling the output of `get_serialized_filters`).  These definitions
are going to be the "single source of truth" for category names and
fields, meaning we can remove a lot of code that attempted to
standardize these or hard-code them.

3. Add an accordion-like component called CategorySection, which will
contain the filter form fields.  Needs some further styling work.

4. Add a `FilterSet` component that handles rendering and managing the
various filters and their values.  Initially only define a few form
fields.  Choice filters and autocomplete filters are slightly more
complicated and will be done separately.