Re-work use of IncidentFilter so it is not required at load time
This commit changes how we are using `IncidentFilter` in

1. The statistics validator
2. The set of choices for the category-based filter settings field

So that we are not forced to import and run the `IncidentFilter` code
when the site itself and its models are initialized, since this
prevents us from initializing or using *other* models in the incident
filter code in a straightforward manner (since you cannot import
models before they've been configured).

I feel that this change will make it easier to write incident filter
code without running into circular import problems or situations where
the site cannot start-up.

The cost is mostly that the set of filter choices is hard-coded
instead of dynamically generated.  I think this is fine, since the
filter choices are not all that complicated, don't change very often,
and I've added a test to make sure they match the "expected" value
based on what the `IncidentFilter` determines they should be from the
incident page model.