Add queryset method to aggregate latest charge updates for incidents
I don't love the way this is implemented, but I'm not sure of an
easier way to do this while still cooperating with the rest of the
`Filter` class framework.

In SQL, it's possible to perform a left join against a subquery,
resulting in something like this:

```sql
SELECT page_ptr_id, array_agg(subquery.most_recent_status)
FROM incident_incidentpage LEFT JOIN (
    SELECT DISTINCT ON (incident_charge_id)
           ic.incident_page_id, ic.charge_id, ic.status AS initial_status, ic.date as initial_date, cu.status AS most_recent_status, cu.date as most_recent_date,
           CASE cu.date < ic.date WHEN true THEN ic.status ELSE coalesce(cu.status, ic.status) END as current_status
    FROM incident_incidentcharge ic
    LEFT JOIN incident_chargeupdate cu ON ic.id = cu.incident_charge_id
    ORDER BY incident_charge_id, cu.date DESC
) subquery ON page_ptr_id = subquery.incident_page_id
GROUP BY page_ptr_id;
```

Which results in (partial example):

```
 page_ptr_id │      array_agg
─────────────┼─────────────────────
          89 │ {NULL}
          90 │ {UNKNOWN,CONVICTED}
          91 │ {NULL}
(3 rows)
```

But from what I can tell (and I tried, here), this SQL is not possible
to represent using normal queryset operations, and while it might be
possible to perform using Django's raw SQL support, I suspect that the
output from those functions would not compose with the rest of the
`Filter` classes.

So I feel like I'm in a little bit of a bind here.  I feel like if we
want more filters like this, it might be worth investigating how to
build the filters on top of a different base, such as a materialized
view.