pip: Do not install dependencies
In theory, pip-tools figures out the dependencies for us and puts what
is needed into the `*requirements.txt` files.  Adding this flag fixes
a bug where sometimes the `--require-hashes` check would fail when
dependencies were followed by pip that were not specified in the
requirements file, and thus were not hashed.