Separate querystring parsing code into its own library
The code to parse the querystring into an object representing the
filter parameters was getting a little unwieldy and depended too
much (IMO) on the "filter definitions" (`filterDefs` -- determined by
the back-end function `get_serialized_filters()`).  This change moves
that code into a separate library and removes that dependency.  The
front-end ought to know how to parse the querystring without input
from the server.

The code itself is a set of names, corresponding to fields in the
querystring, and an associated function for how to extract a value
from the querystring and return a value for storing in the filter
parameters.

Ideally, I'd also like to put code here for the reverse operation:
encoding a querystring from the filter parameters.

One additional note: to reduce the complexity of the code, I've
renamed the `filterCategories` parameters field to just `categories`
so that it matches the querystring field for the categories filter.
This is bringing that filter into line with how the rest of them work.