Model changes: LegalOrder, LegalOrderUpdate, and target field
Here I'm adding the new LegalOrder model, any number of which can
belong to an Incident Page.  Each legal order has any number of
updates.

This is basically the same pattern as the Charges (also in
inlines.py), except with slightly different data on the "base" model.

I've also added a `legal_order_target` field on the Incident Page
model.