Add `categories=<title of category>` to some incident detail links
This commit adds a `categories` parameter to the querystring of some
links on the incident detail page, including all of the links related
to the new legal order metadata.  The reason for this is the incident
index page filter sidebar will begin with the category widget in
"open" or "active" mode if the corresponding category is given in the
querystring.  This makes it easier to find the currently active filter
controls after following a link from an incident card or the incident
detail page.

I've also removed the obsolete entries of `dropped_charges` and
`current_charges`, since these are no longer needed and would
otherwise have to be updated here.