Change legal order update `date` field to optional
Due to the imperfections of the real world, we need legal order
updates to have optional dates.  I think the best UX here is for
updates to be manually orderable, with the last one always being the
"most recent" (even if its date disagrees with that).  I'm told that
the "most recent" status for our real world data _does_ almost always
have a date attached, but the data model needs to account for it
anyway.

This makes minor changes to how the "most recent status" of a
particular legal order is queried.