Enable wide, banner-style special blog cards
This creates a new type of card I'm just calling "special blog card"
here that extends across the full width of the container and uses the
teaser image for the post as a background.  I've tried to stick to the
design as much as possible, though some parameters may need to be
adjusted based on making the styles accomodate the possible images
that will be used.