Enhance tree map chart with "branches" feature
I've added dynamic, selectable "branches" to the tree map.  Previously
this is what was called "categories" but now categories are merely one
possibility for what the tree's branches are.

The admin can select what they want to use as branches in the
`group_by` field on the tree map block.  This provides to the JS
component two properties: the set of branch names, and the name of the
field in the dataset that contains those values.  There is a little
bit of finessing happening in the branch names: sometimes they're
obtained via a URL and sometimes they're passed directly in as an
array-- this is due to how each of the possibilities is stored on the
back end (though it's probably possible to simplify this by passing
everything as an array I was worried the values might get very large).

For the most part the actual tree map chart component is unchanged.
It already basically handles different types of branches, it just
needed a little modification in what data is being passed into it.