Add ChartSnapshot model
Here, I'm adding a model to store/cache the output of the
pregeneration service.  It's called `ChartSnapshot` and it records
calls made to the service, what the service returned, and what time
that happened.

I've included very basic logic for knowing when the cache is too
old (if it was last generated over 24 hours ago).  I've also put a
constraint on the model that it must have at least one value for
`chart_svg` or `chart_image` to be able to be saved, meaning if the
service didn't return anything, we don't have a valid snapshot.