Augment block-related StructValue classes to access chart snapshots
When rendering a template and charts are involved, we have
`StructValue` instances for each of them.  So I think it makes sense
we ought to be able to invoke the snapshot model layer from those
StructValues, meaning we can place the snapshots into our templates
with something like:

```
{{ chart_block.png_snapshot_url }}
```

Or

```
{{ chart_block.svg_snapshot }}
```

In the case of the former, we could even expose the underlying
`ImageField` on `ChartSnapshot` and enable using it with the Wagtail
`image` tag, e.g.:

```
{% image chart_block.png_snapshot width-400 as tmp_snapshot %}
```

Which I think is pretty nice and flexible!