Attempt to decrease repetition in docker compose
And also ensure that we have access to all the settings we need while
_building_ the production container.  The settings `DJANGO_CHART_HOST`
and `DJANGO_CHART_PORT` must be configured when running
`collectstatic`, which the production dockerfile uses in the build
process.  So we have to have these values there.  I've tried to limit
how much duplicatation we have of these by separating them out into an
`x-` extension field (see here
https://docs.docker.com/compose/compose-file/compose-file-v3/#extension-fields)
and it seems to be working all right.