Manually apply teaser images to help with devdata tests
Had some problems with flickering tests locally, this is maybe
helping.  The problem was that the `Iterator` (a factory boy
declaration) was keeping its items in memory across invocations of
`createdevdata` and the old set of items were no longer in the
database after the first time the command completed.  Then, on
subsequent tests that used that command, the incident pages were being
created with no-longer-existing custom images, which caused a
validation error.  This seemed to fix it for me, but keeping an eye on
this in general.