Only run npm_audit and safety_check jobs on develop
This is intended to stop these from running on non-deployment pull
requests, and only trigger them when we are merging from `develop`
into `prod`.

Also remove the check on package-lock.json.  This was needed to
prevent npm audit from running on PRs when the node packages were not
updated.  But now we are not running it on _any_ PRs, so we can remove
that condition.  It ought to always run when we want it run.