Add node/webpack healthcheck
This check parses the `webpack-stats.json` file (with the program `jq`
that has to be installed in the container) and checks if its `status`
key is one of the 3 things we expect: compiling, error, or done.  If
this file exists and is reporting one of those statuses, it's healthy
and ready for django to start using it.