Add health check for django container
This allows us to use `docker compose up --wait` in CI or for local
development as an alternative to `docker compose up --detach`.  Though
it does continuously poll the `/health/ok/` endpoint, which I guess
some may not like?  But I've silenced the logging of that, which helps
reduce the noise.