Check for both possible places `attribution` might be
If we are passed an Image instance, we will have the attribution in
`image.attribution`.

If we are passed a StreamField instance, we will have the attribution
in `image.0.value.attribution`.

Whichever one of these exists (and both might be blank), that should
be used for the the actual rendered attribution text.