Enable overridable summary names for boolean fields
Defines 2 overridable fields on BooleanFilters: present/absent summary
name.  This is the name that will appear in the filter summary section
when the filter is active, for the "present" (True) case and the
"absent" (False) case, respectively.  The server supplies these values
in the serialized filter set.  There are default values that are fine,
but we can override them (as we can for any filter attributes) if we
want to do something more specific.