Grey-out smarch search bar if no more selections can be made
This change makes it so that if the search bar is in a state where no
more selections can be made, then it will appear as greyed-out.  The
search button will still display the hover effect.  I've added the
text label for the "Clear" button, which can also still be used when
the bar is not selectable.