Ensure generated PNGs are saved
We were running into a problem on Google Cloud Storage where the
operation to create a `CustomImage` model would fail if the
`ImageFile` was not saved before adding them to that
`CustomImage`. This, apparently, is not the case for the default
storage backend.

I've updated the code to save the image (using the `pillow` package,
though there may be other ways to do this) returned by the chart
pregenerator API _before_ it's returned to our wagtail code. This
means I've also moved the code that creates a filename based on the
query parameters into the chart API client code, but moving the code
is all that's been done.  It's the same logic.