Rework pregeneration exception structure, add fallback image support
1. I'm making some more specific exception types related to the chart
pregeneration service, and I'm also moving them to be in a more
centralized location (as they're not all directly related to the "API"
module).
2. I'm adding a check for the `ChartNotAvailable` exception (raised
when `get_or_generate` fails both to get _and_ to generate) at the
points where we expect an image for the chart and putting code there
to use the blog page's `search_image` as a fallback.  This will
prevent the whole page from failing to load if the chart is not
available for whatever reason.