Rework chart pregenerator docker setup for live code reloading
Rather than user the Dockerfile command `COPY` to copy the chart
pregenerator code (in `chart_pregenerator`) and chart rendering
code (in `client`) into the node-chart-pregenerator container at build
time, I've set it up to mount both.  This way we can watch the code
for changes as the development process happens.

I've simplified the "watching" code in dev.js to remove explicitly
invoking `chokidar` because it looks like `esbuild` does this on its
own.

I've also updated the `esbuild` package to the latest version.