Update chart JS components and libraries to allow State filtering
We add a new argument to `filterDatasets` (similar to tags or
categories) that will bet a `Set` of states that we want to filter
against.  Then everywhere that uses that function can pass in the
variable (or nothing).

We then supply this argument by threading it through the system: from
the HTML attributes that initialize the components (where our states
are stored as a JSON list) into the different chart initialization JS
and the components themselves.

We also have to accept it as part of the `options` query that is part
of the pregeneration API.  This code also uses `filterDatasets`.